Changed ID type from int to long

+++ src/cmg/cnsim/engine/transaction/Transaction.java
@@ -69,7 +69,7 @@ public class Transaction {
 	 * Constructor for given ID. Time, value and size must be initialized with setters.
 	 * @param id The id of the transaction.
 	 */
-	public Transaction(int id) {
+	public Transaction(long id) {
 		super();
 		this.setID(id);
 	}
@@ -130,7 +130,7 @@ public class Transaction {
 	 * Set the unique ID of the transaction.
 	 * @param ID The ID of the transaction.
 	 */
-	public void setID(int ID) {
+	public void setID(long ID) {
 	    this.ID = ID;
 	}
